requirements.txt
Django==1.6
Django-Select2==4.2.2
FFVideo==0.0.13
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.2.1
PyYAML==3.10
Pygments==1.6
South==0.8.4
Sphinx==1.2
akismet==0.2.0
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.12
celery==3.1.6
distribute==0.6.24
django-angular==0.5.0
django-braces==1.2.2
django-crispy-forms==1.4.0
django-datetime-widget==0.6
django-debug-toolbar==1.0.1
django-extensions==1.2.5
django-extra-views==0.6.3
django-filter==0.7
django-fluent-comments==0.9.2
django-grappelli==2.5.1
django-guardian==1.1.1
django-tinymce==1.5.2
docutils==0.11
easy-thumbnails==1.4
ipython==1.1.0
kombu==3.0.7
psycopg2==2.5.1
pytz==2013.8
six==1.4.1
sphinx-rtd-theme==0.1.5
sqlparse==0.1.10
wsgiref==0.1.2
